refactor(batcher): extract bundle membership into Batcher.Bundles
Move bundle state and structure (membership, held approvals, link
validation, stack edges, merge ordering) into a dedicated submodule,
following the Divider/Message/Labeler pattern. Bundles never posts
comments and never touches batches; orchestration and messaging stay
in the batcher.

Also restores upstream's run/3 name (byte-identical again) and makes
form_stacked/dissolve return the updated structs they persist.
